+++ tubular/mixins.py
@@ -5,31 +5,38 @@ from typing import TYPE_CHECKING
 import narwhals as nw
 import narwhals.selectors as ncs
 import numpy as np
-import pandas as pd
 
 if TYPE_CHECKING:
-    from narwhals.typing import FrameT
+    import pandas as pd
+    from narhwals.typing import FrameT
 
 
 class CheckNumericMixin:
-    def check_numeric_columns(self, X: pd.DataFrame) -> pd.DataFrame:
+    """
+    Mixin class with methods for numeric transformers
+
+    """
+
+    def classname(self) -> str:
+        """Method that returns the name of the current class when called."""
+
+        return type(self).__name__
+
+    @nw.narwhalify
+    def check_numeric_columns(self, X: FrameT) -> FrameT:
         """Helper function for checking column args are numeric for numeric transformers.
 
         Args:
         ----
-            X (pd.DataFrame): Data containing columns to check.
+            X: Data containing columns to check.
 
         """
-        numeric_column_types = X[self.columns].apply(
-            pd.api.types.is_numeric_dtype,
-            axis=0,
+        non_numeric_columns = list(
+            set(self.columns).difference(set(X.select(ncs.numeric()).columns)),
         )
-
-        if not numeric_column_types.all():
-            non_numeric_columns = list(
-                numeric_column_types.loc[~numeric_column_types].index,
-            )
-
+        # sort as set ordering can be inconsistent
+        non_numeric_columns.sort()
+        if len(non_numeric_columns) > 0:
             msg = f"{self.classname()}: The following columns are not numeric in X; {non_numeric_columns}"
             raise TypeError(msg)
 
@@ -138,7 +145,7 @@ class WeightColumnMixin:
         class attribute, indicates whether transformer has been converted to polars/pandas agnostic narwhals framework
     """
 
-    polars_compatible = False
+    polars_compatible = True
 
     @nw.narwhalify
     def check_weights_column(self, X: FrameT, weights_column: str) -> None:
@@ -161,22 +168,22 @@ class WeightColumnMixin:
             raise ValueError(msg)
 
         # check weight is positive
-        if X[weights_column].min() < 0:
+        if X.select(nw.col(weights_column).min()).item() < 0:
             msg = f"{self.classname()}: weight column must be positive"
             raise ValueError(msg)
 
         # check weight non-null
-        if X[weights_column].is_null().sum() != 0:
+        if X.select(nw.col(weights_column).is_null().sum()).item() != 0:
             msg = f"{self.classname()}: weight column must be non-null"
             raise ValueError(msg)
 
-        # check weight not inf
+        # check weight not inf, not currently a narwhals efficient way to do this
         if np.isinf(X[weights_column].to_numpy()).any():
             msg = f"{self.classname()}: weight column must not contain infinite values."
             raise ValueError(msg)
 
         # check weight not all 0
-        if X[weights_column].sum() == 0:
+        if X.select(nw.col(weights_column).sum()).item() == 0:
             msg = f"{self.classname()}: total sample weights are not greater than 0"
             raise ValueError(msg)